requirements .txt
--- Data Manipulation & Engineering ---

pandas>=2.0.0
numpy>=1.24.0

--- Machine Learning & Evaluation ---

scikit-learn>=1.3.0
joblib>=1.3.0

--- Clustering Utilities ---

kneed>=0.8.5

--- Visualization ---

matplotlib>=3.7.0

--- Logic & Reasoning ---

pyswip>=0.2.10

--- Constraint Satisfaction Problem (CSP) ---

python-constraint>=1.4.0

--- API Interaction (Opzionale, se si usa lo script di fetching) ---

requests>=2.31.0